+++ NEWS.rst
@@ -6,6 +6,8 @@ of every change, see the Git log.
 
 Latest
 ------
+* Minor: Added --skip-internal option which allows the user to skip internal
+  dependencies when resolving.
 * Major: Updated to waf 2.0.22
 * Major: Make resolve symlinks relative.
 * Minor: Added support for git_protocol `ssh://git@`